requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
werkzeug
psutil
python-jose
python-multipart